användare

+++ app.rb
@@ -2,6 +2,27 @@ require 'sinatra'
 require 'sqlite3'
 require 'slim'
 require 'sinatra/reloader'
+require 'bcrypt'
+
+enable :sessions
+
+before do
+    session_id = session[:user_id]
+    if session_id == nil
+        @user = nil
+        return
+    end
+
+    db = connectToDb()
+    user = db.execute("SELECT id, username FROM users WHERE id=?", [session_id])
+    if user.empty?
+        @user = nil
+        return
+    end
+
+    @user = user[0]
+    p @user
+end
 
 
 def connectToDb()
@@ -11,8 +32,8 @@ def connectToDb()
     return db
 end
 
-def getTodoById(db, id)
-    todo = db.execute("SELECT * FROM todos WHERE todos.id=?", id)
+def getTodoById(db, id, user_id)
+    todo = db.execute("SELECT * FROM todos WHERE id=? AND owner_id=?", [id, user_id])
     if !todo
         return nil
     end
@@ -20,13 +41,13 @@ def getTodoById(db, id)
     return todo[0]
 end
 
-def getTodoCategories(db, id)
-    categories = db.execute("SELECT categories.id, categories.name FROM categories INNER JOIN todo_cat_rel ON categories.id = todo_cat_rel.category_id WHERE todo_cat_rel.todo_id = ?", id)
+def getTodoCategories(db, id, user_id)
+    categories = db.execute("SELECT categories.id, categories.name FROM categories INNER JOIN todo_cat_rel ON categories.id = todo_cat_rel.category_id WHERE todo_cat_rel.todo_id=? AND categories.owner_id=?", [id, user_id])
     return categories
 end
 
-def getCategoryById(db, id) 
-    category = db.execute("SELECT * FROM categories WHERE id=?", [id])
+def getCategoryById(db, id, user_id) 
+    category = db.execute("SELECT * FROM categories WHERE id=? AND owner_id=?", [id, user_id])
     if !category
         return nil
     end
@@ -39,23 +60,32 @@ get '/' do
 end
 
 get "/todos" do
+    if @user == nil
+        redirect("/login")
+    end
+    user_id = @user["id"]
+
     db = connectToDb()
     
-    todos = db.execute("SELECT * FROM todos")
+    todos = db.execute("SELECT * FROM todos where owner_id=?", [user_id])
     @todo_categories = {}
     todos.each do |todo|
-      @todo_categories[todo["id"]] = getTodoCategories(db, todo["id"])
-      p @todo_categories[todo["id"]]
+      @todo_categories[todo["id"]] = getTodoCategories(db, todo["id"], user_id)
     end
 
     @done_todos = todos.select { |todo| todo["is_done"] == 1 }
     @undone_todos = todos.select { |todo| todo["is_done"] == 0 }
-    @categories = db.execute("SELECT * FROM categories")
+    @categories = db.execute("SELECT * FROM categories WHERE owner_id=?", [user_id])
 
     slim(:"todos/index")
 end
 
 post "/todos" do
+    if @user == nil
+        redirect("/login")
+    end
+    user_id = @user["id"]
+
     db = connectToDb()
 
     name = params[:name]
@@ -67,13 +97,13 @@ post "/todos" do
     end
 
     category_ids.each do |category_id|
-        category = getCategoryById(db, category_id)
+        category = getCategoryById(db, category_id, user_id)
         if !category
             error(404)
         end
     end
 
-    id = db.execute("INSERT INTO todos (name, description) VALUES (?, ?) RETURNING id", [name, description])[0]["id"]
+    id = db.execute("INSERT INTO todos (name, description, owner_id) VALUES (?, ?, ?) RETURNING id", [name, description, user_id])[0]["id"]
     p id
 
     category_ids.each do |category_id|
@@ -84,6 +114,11 @@ post "/todos" do
 end
 
 get "/todos/:id/edit" do
+    if @user == nil
+        redirect("/login")
+    end
+    user_id = @user["id"]
+
     id = params[:id].to_i
     db = connectToDb()
 
@@ -93,17 +128,22 @@ get "/todos/:id/edit" do
         error(404)
     end
 
-    @categories = db.execute("SELECT * FROM categories")
-    @todo_categories = getTodoCategories(db, @todo["id"]).map {|category| category["id"]}
+    @categories = db.execute("SELECT * FROM categories WHERE owner_id=?", [user_id])
+    @todo_categories = getTodoCategories(db, @todo["id"], user_id).map {|category| category["id"]}
 
     slim(:"todos/edit")
 end
 
 post "/todos/:id/update" do
+    if @user == nil
+        redirect("/login")
+    end
+    user_id = @user["id"]
+
     id = params[:id].to_i
     db = connectToDb()
 
-    @todo = getTodoById(db, id)
+    @todo = getTodoById(db, id, user_id)
     if !@todo
         error(404)
     end
@@ -116,50 +156,65 @@ post "/todos/:id/update" do
     p params.inspect
 
     category_ids.each do |category_id|
-        category = getCategoryById(db, category_id)
+        category = getCategoryById(db, category_id, user_id)
         if !category
             error(404)
         end
     end
 
     if is_done != nil and name == nil
-        db.execute("UPDATE todos SET is_done=? WHERE id=?", [is_done, id])
+        db.execute("UPDATE todos SET is_done=? WHERE id=? AND owner_id=?", [is_done, id, user_id])
     else
-        db.execute("DELETE FROM todo_cat_rel WHERE todo_id = ?", [id])
+        db.execute("DELETE FROM todo_cat_rel WHERE todo_id=?", [id])
 
         category_ids.each do |category_id|
             db.execute("INSERT INTO todo_cat_rel (todo_id, category_id) VALUES (?, ?)", [id, category_id])
         end
 
-        db.execute("UPDATE todos SET name=?, description=?, is_done=? WHERE id=?", [name, description, is_done, id])
+        db.execute("UPDATE todos SET name=?, description=?, is_done=? WHERE id=? and owner_id=?", [name, description, is_done, id, user_id])
     end
 
     redirect("/todos")
 end
 
 post "/todos/:id/delete" do
+    if @user == nil
+        redirect("/login")
+    end
+    user_id = @user["id"]
+
     id = params[:id].to_i
     db = connectToDb()
 
-    @todo = getTodoById(db, id)
+    @todo = getTodoById(db, id, user_id)
     if !@todo
         error(404)
     end
 
-    db.execute("DELETE FROM todos WHERE id=?", id)
+    db.execute("DELETE FROM todos WHERE id=? AND owner_id=?", [id, user_id])
 
     redirect("/todos")
 end
 
 get "/categories" do
+    if @user == nil
+        redirect("/login")
+    end
+    user_id = @user["id"]
+
     db = connectToDb()
 
-    @categories = db.execute("SELECT * FROM categories")
+    @categories = db.execute("SELECT * FROM categories WHERE owner_id=?", [user_id])
 
     slim(:"categories/index")
 end
 
 post "/categories" do
+    if @user == nil
+        redirect("/login")
+    end
+    user_id = @user["id"]
+
     db = connectToDb()
 
     name = params[:name]
@@ -167,40 +222,55 @@ post "/categories" do
         error(400)
     end
 
-    db.execute("INSERT INTO categories (name) VALUES (?)", [name])
+    db.execute("INSERT INTO categories (name, owner_id) VALUES (?, ?)", [name, user_id])
     
     redirect("/categories")
 end
 
 get "/categories/:id/edit" do
+    if @user == nil
+        redirect("/login")
+    end
+    user_id = @user["id"]
+
     db = connectToDb()
     id = params[:id].to_i
 
-    @category = getCategoryById(db, id)
+    @category = getCategoryById(db, id, user_id)
 
     slim(:"categories/edit")
 end
 
 post "/categories/:id/update" do
+    if @user == nil
+        redirect("/login")
+    end
+    user_id = @user["id"]
+
     db = connectToDb()
     id = params[:id].to_i
 
-    category = getCategoryById(db, id)
+    category = getCategoryById(db, id, user_id)
     if !category
         error(404)
     end
 
     name = params[:name]
 
-    db.execute("UPDATE categories SET name=? WHERE id=?", [name, id])
+    db.execute("UPDATE categories SET name=? WHERE id=? AND owner_id=?", [name, id, user_id])
     redirect("/categories")
 end
 
 post "/categories/:id/delete" do
+    if @user == nil
+        redirect("/login")
+    end
+    user_id = @user["id"]
+
     db = connectToDb()
     id = params[:id].to_i
 
-    category = getCategoryById(db, id)
+    category = getCategoryById(db, id, user_id)
     if !category
         error(404)
     end
@@ -209,4 +279,86 @@ post "/categories/:id/delete" do
     db.execute("DELETE FROM categories WHERE id=?", id)
 
     redirect("/categories")
+end
+
+get "/user" do
+    if @user != nil
+        redirect("/todos")
+    end
+
+    slim(:user)
+end
+
+post "/user" do
+    if @user != nil
+        redirect("/todos")
+    end
+
+    username = params["username"]
+    pass = params["pass"]
+    pass_con = params["pass_conf"]
+
+    p pass
+    p pass_con
+
+    db = connectToDb()
+    existing_user = db.execute("SELECT id FROM users WHERE username=?", username)
+    if existing_user.empty?
+        if pass == pass_con
+            pass_digest = BCrypt::Password.create(pass)
+            row = db.execute("INSERT INTO users (username, pass_digest) VALUES (?, ?) RETURNING id", [username, pass_digest])
+            user_id = row[0]["id"]
+
+            session[:user_id] = user_id
+            
+            redirect("/todos")
+        else
+            redirect("/user")
+        end
+    else
+        redirect("/login")
+    end
+end
+
+get "/login" do
+    if @user != nil
+        redirect("/todos")
+    end
+
+    slim(:login)
+end
+
+post "/login" do
+    if @user != nil
+        redirect("/todos")
+    end
+
+    username = params["username"]
+    pass = params["pass"]
+
+    db = connectToDb()
+    existing_user = db.execute("SELECT id, pass_digest FROM users WHERE username=?", username)
+    if existing_user.empty?
+        redirect("/login")
+    end
+
+    user_id = existing_user[0]["id"]
+    pass_digest = existing_user[0]["pass_digest"]
+
+    if BCrypt::Password.new(pass_digest) == pass 
+        session[:user_id] = user_id
+        redirect("/todos")
+    end
+
+    redirect("/login")
+end
+
+post "/logout" do
+    if !@user then
+        error(401)
+    end
+
+    session[:user_id] = nil
+
+    redirect("/login")
 end
